% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_bis}
\alias{load_data_bis}
\title{This funciton takes the file location and convert raw data into CNV class for downstream analysis.}
\usage{
load_data_bis(fileloc)
}
\value{
a CNV class containing raw data information
}
\description{
This funciton takes the file location and convert raw data into CNV class for downstream analysis.
}
