% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.function.R
\name{makeBatch}
\alias{makeBatch}
\title{Building a Model with Top Ten Features}
\usage{
makeBatch(data = data, keyMap_std = "default", calib,
  locMap_plate_96 = "rivur")
}
\arguments{
\item{data}{a dataframe of the data that is pased}

\item{keyMap_std}{a dataframe containing the location information of standards used.}

\item{calib}{a dataframe containing calibrator information}
}
\value{
a list of results for data summary, std summary, differnce between estimated and rounded copy numbers and sample location.
}
\description{
This funciton develops a prediction algorithm based on the top10 features in 'x' that are most predictive of 'y'.
}
\details{
This function runs a multivaritate regression of y on each predictor in x and calculates a p-value indicating the significance of the association. The final set of 10 predictors is taken from the 10 smallest p-values.
}
\seealso{
\code lm
}
\author{
Dong Liang
}
