% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData_CaseCtl.R
\name{makeData_CaseCtl}
\alias{makeData_CaseCtl}
\title{Function currently under test for case control plate data}
\usage{
makeData_CaseCtl(data_raw, keyMap_std = "default", zeroCopy = NULL)
}
\description{
Function currently under test for case control plate data
}
