% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.function.R
\name{makeDataBatch}
\alias{makeDataBatch}
\title{This funciton takes the file location and convert raw data into CNV class for downstream analysis.}
\usage{
makeDataBatch(file_list, stdKeyMap, calib, locMap_plate_96 = "rivur",
  BadData_remove = TRUE)
}
\value{
a data bundle containing all the raw data information
}
\description{
This funciton takes the file location and convert raw data into CNV class for downstream analysis.
}
