% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat}
\alias{stat}
\title{This funciton perform statistical analysis.}
\usage{
stat(data, gene, fill, chisq = F)
}
\arguments{
\item{data}{a dataframe of the data that is pased}

\item{gene}{a character input stating the dependent variable to be used for statistical analysis}

\item{fill}{a character input stating the independent variable to be used for statistical analysis}
}
\value{
a list of statistical test results including p values, number of obs, method used, and test itself as a class.
}
\description{
This funciton perform statistical analysis.
}
\details{
This function runs a multivaritate regression of y on each predictor in x and calculates a p-value indicating the significance of the association. The final set of 10 predictors is taken from the 10 smallest p-values.
}
\seealso{
\code lm
}
\author{
Dong Liang
}
